Add support for running wdspec tests in Chrome.
Most of the infrastructure for running wdspec tests is independent of
selenium vs marionette vs etc since the tests themselves define the
protocol. So most of the implementation is moved into the base module
and each executor just defines how it starts the WebDriver binary.

The chrome support is then a matter of starting the right server and
hooking up the glue in browsers/chrome.py. We set the w3c capability
for maximum spec compatibility.
